tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/send/009.html

<!doctype html>
<title>WebSockets: send('')</title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
async_test(function(t){
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/empty-message');
  ws.onopen = t.step_func(function(e) {
    ws.send('');
  })
  ws.onmessage = t.step_func(function(e) {
    assert_equals(e.data, 'pass');
    ws.close();
  });
  ws.onclose = t.step_func(function(e) {
    ws.onclose = t.step_func(function() {assert_unreached()});
    setTimeout(function() {t.done()}, 50);
  });
  ws.onerror = t.step_func(function() {assert_unreached()});
});
</script>
